Factor CodeMirror 6 editor out into separate package
