+++ .gitignore
@@ -399,23 +399,7 @@ packages/app-mobile/components/NoteBodyViewer/NoteBodyViewer.js
 packages/app-mobile/components/NoteBodyViewer/hooks/useOnMessage.js
 packages/app-mobile/components/NoteBodyViewer/hooks/useOnResourceLongPress.js
 packages/app-mobile/components/NoteBodyViewer/hooks/useSource.js
-packages/app-mobile/components/NoteEditor/CodeMirror/CodeMirror.test.js
 packages/app-mobile/components/NoteEditor/CodeMirror/CodeMirror.js
-packages/app-mobile/components/NoteEditor/CodeMirror/decoratorExtension.js
-packages/app-mobile/components/NoteEditor/CodeMirror/markdownCommands.bulletedVsChecklist.test.js
-packages/app-mobile/components/NoteEditor/CodeMirror/markdownCommands.test.js
-packages/app-mobile/components/NoteEditor/CodeMirror/markdownCommands.toggleList.test.js
-packages/app-mobile/components/NoteEditor/CodeMirror/markdownCommands.js
-packages/app-mobile/components/NoteEditor/CodeMirror/markdownMathParser.test.js
-packages/app-mobile/components/NoteEditor/CodeMirror/markdownMathParser.js
-packages/app-mobile/components/NoteEditor/CodeMirror/markdownReformatter.test.js
-packages/app-mobile/components/NoteEditor/CodeMirror/markdownReformatter.js
-packages/app-mobile/components/NoteEditor/CodeMirror/syntaxHighlightingLanguages.js
-packages/app-mobile/components/NoteEditor/CodeMirror/testUtil/createEditor.js
-packages/app-mobile/components/NoteEditor/CodeMirror/testUtil/forceFullParse.js
-packages/app-mobile/components/NoteEditor/CodeMirror/testUtil/loadLanguages.js
-packages/app-mobile/components/NoteEditor/CodeMirror/theme.js
-packages/app-mobile/components/NoteEditor/CodeMirror/types.js
 packages/app-mobile/components/NoteEditor/CodeMirror/webviewLogger.js
 packages/app-mobile/components/NoteEditor/EditLinkDialog.js
 packages/app-mobile/components/NoteEditor/MarkdownToolbar/MarkdownToolbar.js
@@ -428,7 +412,6 @@ packages/app-mobile/components/NoteEditor/MarkdownToolbar/types.js
 packages/app-mobile/components/NoteEditor/NoteEditor.test.js
 packages/app-mobile/components/NoteEditor/NoteEditor.js
 packages/app-mobile/components/NoteEditor/SearchPanel.js
-packages/app-mobile/components/NoteEditor/SelectionFormatting.js
 packages/app-mobile/components/NoteEditor/types.js
 packages/app-mobile/components/NoteList.js
 packages/app-mobile/components/ProfileSwitcher/ProfileEditor.js
@@ -478,6 +461,26 @@ packages/app-mobile/utils/fs-driver-rn.js
 packages/app-mobile/utils/setupNotifications.js
 packages/app-mobile/utils/shareHandler.js
 packages/app-mobile/utils/types.js
+packages/editor/CodeMirror/computeSelectionFormatting.js
+packages/editor/CodeMirror/createEditor.test.js
+packages/editor/CodeMirror/createEditor.js
+packages/editor/CodeMirror/decoratorExtension.js
+packages/editor/CodeMirror/markdownCommands.bulletedVsChecklist.test.js
+packages/editor/CodeMirror/markdownCommands.test.js
+packages/editor/CodeMirror/markdownCommands.toggleList.test.js
+packages/editor/CodeMirror/markdownCommands.js
+packages/editor/CodeMirror/markdownMathParser.test.js
+packages/editor/CodeMirror/markdownMathParser.js
+packages/editor/CodeMirror/markdownReformatter.test.js
+packages/editor/CodeMirror/markdownReformatter.js
+packages/editor/CodeMirror/syntaxHighlightingLanguages.js
+packages/editor/CodeMirror/testUtil/createEditor.js
+packages/editor/CodeMirror/testUtil/forceFullParse.js
+packages/editor/CodeMirror/testUtil/loadLanguages.js
+packages/editor/CodeMirror/theme.js
+packages/editor/SelectionFormatting.js
+packages/editor/events.js
+packages/editor/types.js
 packages/fork-htmlparser2/src/CollectingHandler.js
 packages/fork-htmlparser2/src/FeedHandler.spec.js
 packages/fork-htmlparser2/src/FeedHandler.js
