+++ packages/app-mobile/package.json
@@ -18,7 +18,8 @@
     "postinstall": "jetify && yarn run build"
   },
   "dependencies": {
-    "@bam.tech/react-native-image-resizer": "3.0.7",
+    "@bam.tech/react-native-image-resizer": "3.0.5",
+    "@joplin/editor": "~2.13",
     "@joplin/lib": "~2.13",
     "@joplin/react-native-alarm-notification": "~2.13",
     "@joplin/react-native-saf-x": "~2.13",
@@ -86,19 +87,6 @@
     "@babel/core": "7.20.2",
     "@babel/preset-env": "7.20.2",
     "@babel/runtime": "7.20.0",
-    "@codemirror/commands": "6.2.2",
-    "@codemirror/lang-cpp": "6.0.2",
-    "@codemirror/lang-html": "6.4.3",
-    "@codemirror/lang-java": "6.0.1",
-    "@codemirror/lang-javascript": "6.1.5",
-    "@codemirror/lang-markdown": "6.1.0",
-    "@codemirror/lang-php": "6.0.1",
-    "@codemirror/lang-rust": "6.0.1",
-    "@codemirror/language": "6.6.0",
-    "@codemirror/legacy-modes": "6.3.2",
-    "@codemirror/search": "6.3.0",
-    "@codemirror/state": "6.2.0",
-    "@codemirror/view": "6.9.3",
     "@joplin/tools": "~2.13",
     "@lezer/highlight": "1.1.4",
     "@testing-library/jest-native": "5.4.3",
