+++ packages/editor/CodeMirror/syntaxHighlightingLanguages.ts
