+++ packages/editor/CodeMirror/testUtil/forceFullParse.ts
