+++ packages/editor/package.json
@@ -1,0 +1,45 @@
+{
+	"name": "@joplin/editor",
+	"version": "2.13.0",
+	"description": "Web-based markdown editor",
+	"private": true,
+	"scripts": {
+		"tsc": "tsc --project tsconfig.json",
+		"watch": "tsc --watch --preserveWatchOutput --project tsconfig.json",
+		"postinstall": "yarn run tsc",
+		"test": "jest",
+		"test-ci": "yarn test"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/laurent22/joplin.git"
+	},
+	"devDependencies": {
+		"@joplin/lib": "~2.13",
+		"@joplin/tools": "~2.13",
+		"@testing-library/react-hooks": "8.0.1",
+		"@types/jest": "29.5.3",
+		"@types/react": "18.0.24",
+		"@types/react-redux": "7.1.25",
+		"@types/styled-components": "5.1.26",
+		"jest": "29.5.0",
+		"jest-environment-jsdom": "29.5.0",
+		"ts-jest": "29.1.1",
+		"typescript": "5.1.3"
+	},
+	"dependencies": {
+		"@codemirror/commands": "6.2.2",
+		"@codemirror/lang-cpp": "6.0.2",
+		"@codemirror/lang-html": "6.4.3",
+		"@codemirror/lang-java": "6.0.1",
+		"@codemirror/lang-javascript": "6.1.5",
+		"@codemirror/lang-markdown": "6.1.0",
+		"@codemirror/lang-php": "6.0.1",
+		"@codemirror/lang-rust": "6.0.1",
+		"@codemirror/language": "6.6.0",
+		"@codemirror/legacy-modes": "6.3.2",
+		"@codemirror/search": "6.3.0",
+		"@codemirror/state": "6.2.0",
+		"@codemirror/view": "6.9.3"
+	}
+}
